requirements.txt
pandas
numpy
altair
streamlit-aggrid
